// honey-paradise-client/src/shared/types/server/createUserDto.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Honey Paradise
 * Honey Paradise API
 * OpenAPI spec version: 0.0.2
 */
import type { CreateUserDtoGender } from './createUserDtoGender';

export interface CreateUserDto {
  email: string;
  password: string;
  username?: string;
  gender?: CreateUserDtoGender;
  birthdate?: string;
}